Refactor client low-level conn module
