+++ justfile
@@ -38,6 +38,7 @@ clippy:
 examples:
     cargo +{{rust}} check --examples --all-features
 
+alias d := docs
 # Build documentation
 docs:
     cargo +{{rust}} doc --all-features --no-deps
