+++ src/discovery/transport.rs
@@ -11,7 +11,7 @@ use hyper::Uri;
 use super::ConnectionError;
 use super::ServiceRegistry;
 use crate::client::conn::Stream as ClientStream;
-use crate::client::TransportStream;
+use crate::client::conn::TransportStream;
 
 pub use builder::TransportBuilder;
